pyproject.toml
[tool.poetry]
name = "screenplay-dreamer"
version = "0.1.0"
description = "Turn your photos into cinematic screenplays using AI"
authors = ["Wietse Venema"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
google-genai = "^0.6.0"
langgraph = "^0.2.68"
google-cloud-firestore = "^2.20.0"
google-cloud-storage = "^2.19.0"
pillow = "^11.1.0"
pillow-heif = "^0.21.0"
google-auth = "^2.38.0"
pyjwt = "^2.10.1"
pydantic = "^2.10.6"
pydantic-settings = "^2.7.1"

[tool.poetry.dev-dependencies]
black = "^25.1.0"
pre-commit = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
